+++ actix-service/src/apply.rs
@@ -208,14 +208,12 @@ where
 
 #[cfg(test)]
 mod tests {
+    use core::future::{ready, Ready};
+
     use futures_util::future::lazy;
 
     use super::*;
-    use crate::{
-        ok,
-        pipeline::{pipeline, pipeline_factory},
-        Ready,
-    };
+    use crate::pipeline::{pipeline, pipeline_factory};
 
     #[derive(Clone)]
     struct Srv;
@@ -228,7 +226,7 @@ mod tests {
         crate::always_ready!();
 
         fn call(&self, _: ()) -> Self::Future {
-            ok(())
+            ready(Ok(()))
         }
     }
 
@@ -252,7 +250,7 @@ mod tests {
     #[actix_rt::test]
     async fn test_new_service() {
         let new_srv = pipeline_factory(apply_fn_factory(
-            || ok::<_, ()>(Srv),
+            || ready(Ok::<_, ()>(Srv)),
             |req: &'static str, srv| {
                 let fut = srv.call(());
                 async move {