+++ actix-service/CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Minimum supported Rust version (MSRV) is now 1.88.
+- internal: Use `core::future::{ready, Ready}` instead of hand-crafted one
 
 ## 2.0.3