+++ actix-utils/CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Minimum supported Rust version (MSRV) is now 1.88.
+- Deprecate `crate::ready::*` in favor of `core::future::{ready, Ready}`
 
 ## 3.0.1